ungroup is sent to the single speaker instead of the master
commands in AVTransport are sent to the group master by default. this is correct for most commands, like play for example. sending the ungroup command to the master results in removing the master from the group instead of the referenced speaker. the ungroup command must be sent to the referenced speaker, not the master.